finished add to cart functionality but the update and delete has bug have to fix this later. day 54 cart functionality completed

+++ ECOMMERCE FRONTEND/src/pages/auth/cart/Cart.tsx
@@ -1,6 +1,6 @@
 import Footer from "../../../globals/components/footer/Footer";
 import Navbar from "../../../globals/components/navbar/Navbar";
-import { deleteCartItem } from "../../../store/cartSlice";
+import { deleteCartItem, updateCartItem } from "../../../store/cartSlice";
 import { useAppDispatch, useAppSelector } from "../../../store/hooks";
 
 const Cart = () => {
@@ -9,10 +9,21 @@ const Cart = () => {
   const dispatch = useAppDispatch();
 
   const handleDelete = (productId: string) => {
-    dispatch(deleteCartItem(productId));
-    
-  };
 
+    if (productId) {
+        dispatch(deleteCartItem(productId));
+    } else {
+        console.error("Product ID is undefined!!!!!!");
+    }
+};
+
+
+const handleUpdate = (productId:string, quantity:number)=>{
+  dispatch(updateCartItem(productId, quantity))
+}
+
+const totalItemInCarts = items.reduce((total, item)=>item?.quantity + total, 0)//0 is accumulator and the initial value of total is 0.
+const totalPriceInCarts = items.reduce((total, item)=>item?.Product?.productPrice * item?.quantity + total, 0)
   
   return (
     <>
@@ -23,50 +34,44 @@ const Cart = () => {
           <div className="rounded-lg md:w-2/3">
             {items.length > 0 &&
               items.map((item, index) => {
-                
                 return (
                   <div
-                    key={item.Product.id || index}
+                    key={item?.Product?.id || index}
                     className="justify-between mb-6 rounded-lg bg-white p-1 shadow-md sm:flex sm:justify-start "
                   >
                     <img
                       className="w-full h-48 object-cover rounded-xl"
-                      src={`http://localhost:3000/uploads/${item.Product.productImageUrl}`}
+                      src={`http://localhost:3000/uploads/${item?.Product?.productImageUrl}`}
                     />
                     <div className="sm:ml-4 sm:flex sm:w-full sm:justify-between">
                       <div className="mt-5 sm:mt-0">
                         <h2 className="text-lg font-bold text-gray-900">
-                          {item.Product.productName}
+                          {item?.Product?.productName}
                         </h2>
                         <p className="mt-1 text-xs text-blue-800 font-bold">
-                          {item.Product?.category?.categoryName}
+                          {item?.Product?.category?.categoryName}
                         </p>
                         <p className="mt-1 text-xs text-gray-700">
-                          {item.Product?.productDescription}
+                          {item?.Product?.productDescription}
                         </p>
                       </div>
                       <div className="mt-4 flex justify-between sm:space-y-6 sm:mt-0 sm:block sm:space-x-6">
                         <div className="flex items-center border-gray-100">
-                          <span className="cursor-pointer rounded-l bg-gray-700 py-1 px-3.5 duration-100 hover:bg-emerald-500 hover:text-blue-50">
-                            {" "}
-                            -{" "}
-                          </span>
+                          <span className="cursor-pointer rounded-l bg-gray-700 py-1 px-3.5 duration-100 hover:bg-emerald-500 hover:text-blue-50" onClick={()=>handleUpdate(item?.Product?.id, item?.quantity - 1)}>-</span>
                           <input
                             className="h-8 w-8 border bg-black text-center text-xs outline-none"
                             type="number"
-                            value={item.Quantity}
+                            value={item?.quantity}
                             min="1"
                           />
-                          <span className="cursor-pointer rounded-r bg-gray-700 py-1 px-3 duration-100 hover:bg-emerald-500 hover:text-blue-50">
-                            {" "}
-                            +{" "}
+                          <span className="cursor-pointer rounded-r bg-gray-700 py-1 px-3 duration-100 hover:bg-emerald-500 hover:text-blue-50" onClick={()=>handleUpdate(item?.Product?.id, item?.quantity + 1)}>+
                           </span>
                         </div>
                         <div className="flex items-center space-x-4">
                           <p className="text-sm text-green-700 font-bold">
-                            Price: ${item.Product?.productPrice}
+                            Price: ${item?.Product?.productPrice}
                           </p>
-                          <button onClick={() => handleDelete(item.Product.id)} className="text-white mt-6 w-full rounded-2xl bg-red-500 px-2 py-1 font-sm hover:bg-red-600">Remove</button>
+                          <button onClick={() => handleDelete(item?.Product?.id)} className="text-white mt-6 w-full rounded-2xl bg-red-500 px-2 py-1 font-sm hover:bg-red-600">Remove</button>
                         </div>
                       </div>
                     </div>
@@ -76,9 +81,13 @@ const Cart = () => {
           </div>
 
           <div className="mt-6 h-full rounded-lg border bg-white p-6 shadow-md md:mt-0 md:w-1/3">
+          <div className="mb-2 flex justify-between">
+              <p className="text-gray-700">Total Items:</p>
+              <p className="text-gray-700">{totalItemInCarts}</p>
+            </div>
             <div className="mb-2 flex justify-between">
               <p className="text-gray-700">Subtotal</p>
-              <p className="text-gray-700">$129.99</p>
+              <p className="text-gray-700">${totalPriceInCarts}</p>
             </div>
             <div className="flex justify-between">
               <p className="text-gray-700">Shipping</p>
@@ -89,7 +98,7 @@ const Cart = () => {
               <p className="text-lg font-bold text-green-700">Total:</p>
               <div className="">
                 <p className="mb-1 text-lg font-bold text-green-800">
-                  $134.98 USD
+                  ${totalPriceInCarts + 4.99}
                 </p>
                 <p className="text-sm text-gray-700">including VAT</p>
               </div>
